covert_side_channels: mention concurrency in execution history
The word "inheritance" evokes something that happened in the past.
Try to clarify that the execution history spans the whole
microarchitecture, consisting of many complex concurrently execution
units.

fence.time cannot do anything against concurrency, so we'll have to
mention concurrency in generic recommendations to implementers, because
software is powerless.

Signed-off-by: Radim Krčmář <rkrcmar@ventanamicro.com>